Complete DebugExercise1.

+++ lab2/DebugPractice/DebugExercise1.java
@@ -3,25 +3,25 @@
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+    public static int divideThenRound(double top, double bottom) {
+        double quotient = top / bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
 
     public static void main(String[] args) {
-        int t = 10;
-        int b = 2;
+        double t = 10;
+        double b = 2;
         int result = divideThenRound(t, b);
         System.out.println("round(" + t + "/" + b + ")=" + result);
 
-        int t2 = 9;
-        int b2 = 4;
+        double t2 = 9;
+        double b2 = 4;
         int result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
 
-        int t3 = 3;
-        int b3 = 4;
+        double t3 = 3;
+        double b3 = 4;
         int result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
     }
